Update project 24032023

+++ src/main/resources/schema.sql
@@ -2,14 +2,14 @@ DROP TABLE IF EXISTS tasks;
 DROP TABLE IF EXISTS users;
 
 CREATE TABLE tasks (
-    id BIGINT PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name VARCHAR(255),
     is_done BOOLEAN,
     deadline TIMESTAMP
 );
 
 CREATE TABLE users (
-    id BIGINT PRIMARY KEY,
+    id SERIAL  PRIMARY KEY,
     first_name VARCHAR(255),
     last_name VARCHAR(255),
     email VARCHAR(255),
